Use uint8_t for WiFi SSID length

+++ components/ocs_net/ap_network_config.h
@@ -76,7 +76,7 @@ private:
     static constexpr const char* channel_key_ = "channel";
     static constexpr const char* max_conn_key_ = "max_conn";
 
-    static constexpr unsigned max_ssid_len_ = 31;
+    static constexpr uint8_t max_ssid_len_ = 31;
     static constexpr uint8_t default_password_len_ = 12;
 
     storage::IStorage& storage_;
